Miscellaneous localization fixes
• Replaced ":" with "=" as a key/value separator, with the first used only
for two keys, actionNotifyReadOnlyMsg and actionReplyReadOnlyMsg, in a
few localizations, compared to the "=" separator used everywhere else.

• Added missing trailing newline at EOF for a few files

• Fixed unescaped newline for value of key fontSize

• Fixed unescaped newline in the value of key aboutMessageZD

• Useless trailing newline removed for key im.MaxUsers in the Polish
localization

• Removed trailing newline inconsistent with all other localizations.

• Fixed error for key basicClientLoginNotice in Arabic localization:
the whole definition of the key bcc, "bcc = Bcc", was duplicated as
the trailing end of the previous key, basicClientLoginNotice.
Removed spurious string.

• Removed useless trailing newline absent everywhere else for key disableOfflineDetails

• Fixed garbled definition of the Etc/GMT+12 timezone, with key missing.

• Fixed unescaped newline for key confirmEmptyTrashFolder in the Thai localization

• Fixed wrong value for key switchIncludeAndFormat in the French localization:
the whole key/value definition for the key switchIncludeWarning was
duplicated in the value of the key switchIncludeAndFormat.
Probable Copy&Paste error. Value fixed.